// src/types/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Category = {
  __typename?: 'Category';
  id: Scalars['Int'];
  image_url: Scalars['String'];
  name: Scalars['String'];
};

export type CreateCategoryInput = {
  image_url: Scalars['String'];
  name: Scalars['String'];
};

export type CreateItemInput = {
  category_id: Scalars['Int'];
  image_url: Scalars['String'];
  name: Scalars['String'];
  price: Scalars['Float'];
  url: Scalars['String'];
};

export type Item = {
  __typename?: 'Item';
  id: Scalars['Int'];
  image_url: Scalars['String'];
  name: Scalars['String'];
  price: Scalars['Float'];
  url: Scalars['String'];
};

export type Mutation = {
  __typename?: 'Mutation';
  createCategory?: Maybe<Category>;
  createItem?: Maybe<Item>;
};


export type MutationCreateCategoryArgs = {
  input?: InputMaybe<CreateCategoryInput>;
};


export type MutationCreateItemArgs = {
  input?: InputMaybe<CreateItemInput>;
};

export type Query = {
  __typename?: 'Query';
  getCategories?: Maybe<Array<Category>>;
  getCategory?: Maybe<Category>;
  getCategoryItems?: Maybe<Array<Item>>;
};


export type QueryGetCategoryArgs = {
  categoryId: Scalars['Int'];
};


export type QueryGetCategoryItemsArgs = {
  categoryId: Scalars['Int'];
};